move askJetpack to get-suggestion-with-stream.js
